Move file-related items into a submodule
