+++ codespan-reporting/examples/term.rs
@@ -1,9 +1,9 @@
 use structopt::StructOpt;
 
 use codespan_reporting::diagnostic::{Diagnostic, Label};
+use codespan_reporting::files::SimpleFiles;
 use codespan_reporting::term::termcolor::StandardStream;
-use codespan_reporting::term::{emit, ColorArg};
-use codespan_reporting::SimpleFiles;
+use codespan_reporting::term::{self, ColorArg};
 
 #[derive(Debug, StructOpt)]
 #[structopt(name = "emit")]
@@ -149,6 +149,6 @@ fn main() {
     let writer = StandardStream::stderr(opts.color.into());
     let config = codespan_reporting::term::Config::default();
     for diagnostic in &diagnostics {
-        emit(&mut writer.lock(), &config, &files, &diagnostic).unwrap();
+        term::emit(&mut writer.lock(), &config, &files, &diagnostic).unwrap();
     }
 }
