+++ codespan-reporting/src/term/views/locus.rs
@@ -1,8 +1,8 @@
 use std::io;
 use termcolor::WriteColor;
 
+use crate::files::Location;
 use crate::term::Config;
-use crate::Location;
 
 /// The 'location focus' of a source code snippet.
 ///
